Fix the version string and release 0.4.1 (#20)
Vault introduced a versioning system for plugins. The plugin must
reports its version. vault-plugin-secrets-exoscale implemented this
versioning in 0.4.0.

It seems like vault requires the version string to start with the "v"
prefix, we didn't do that in "0.4.0".

This PR adds the v prefix and releases a new version of the plugin.

If you installed 0.4.0 you might not be able to unregister it: If you
installed the plugin using `vault plugin register` without specifying
the `-version` flag, vault automatically retrieves the version number
and uses it without validating the "v" prefix. The API call to
unregister plugin is however not capable of removing a versioned plugin
that doesn't have the "v" prefix. If this is your case, you can use
[sys/raw](https://developer.hashicorp.com/vault/api-docs/system/raw) to
remove the bogus version of the plugin:

```
# ⚠️ WARNING ⚠️ 
# Please backup vault and read the `sys/raw` documentation before doing this
# Editing `sys/raw` manually can break your vault cluster permanently

vault delete -format=json /sys/raw/core/plugin-catalog/secret/exoscale/0.4.0
```

If you didn't install 0.4.0, you can upgrade directly to 0.4.1