Fix serialization bug preventing from setting a role as non-renewable (#24)
We set `role.Renewable` to `true` before decoding JSON to make sure that
roles that don't contain that field still keep the old behavior from
before we allowed keys to be non-renewable.

Since this field has `omitempty`, it is omitted when it is set to its
zero value `false`.
This is a bug and prevents the user from making keys non-renewable.